Added user permissions
