+++ internal/whgoxy/discord/attributes.go
@@ -1,12 +1,14 @@
 package discord
 
 type attributes struct {
-	MaxWebhookCount uint `json:"max_webhook_count" bson:"max_webhook_count"`
+	MaxWebhookCount uint   `json:"max_webhook_count" bson:"max_webhook_count"`
+	Permissions     uint64 `json:"permissions" bson:"permissions"`
 }
 
 func NewDefaultAttributes() *attributes {
 	return &attributes{
 		MaxWebhookCount: 20,
+		Permissions:     0,
 	}
 }
 
@@ -15,5 +17,6 @@ func (a *attributes) Repair() (updated bool) {
 		a.MaxWebhookCount = 20
 		updated = true
 	}
+	// Permissions: 0
 	return
 }
